+++ core/Lucy/Store/CompoundFileReader.c
@@ -92,7 +92,7 @@ CFReader_do_open(CompoundFileReader *self, Folder *folder) {
         String *folder_name = IxFileNames_local_part(Folder_Get_Path(folder));
         size_t folder_name_len = Str_Length(folder_name);
 
-        for (uint32_t i = 0, max = Vec_Get_Size(files); i < max; i++) {
+        for (size_t i = 0, max = Vec_Get_Size(files); i < max; i++) {
             String *orig = (String*)Vec_Fetch(files, i);
             if (Str_Starts_With(orig, folder_name)) {
                 Obj *record = Hash_Delete(ivars->records, orig);
@@ -318,7 +318,7 @@ CFReaderDH_Next_IMP(CFReaderDirHandle *self) {
         ivars->tick++;
         if (ivars->tick < (int32_t)Vec_Get_Size(ivars->elems)) {
             String *path = (String*)CERTIFY(
-                                Vec_Fetch(ivars->elems, ivars->tick), STRING);
+                                Vec_Fetch(ivars->elems, (size_t)ivars->tick), STRING);
             DECREF(ivars->entry);
             ivars->entry = (String*)INCREF(path);
             return true;
@@ -335,7 +335,7 @@ bool
 CFReaderDH_Entry_Is_Dir_IMP(CFReaderDirHandle *self) {
     CFReaderDirHandleIVARS *const ivars = CFReaderDH_IVARS(self);
     if (ivars->elems) {
-        String *name = (String*)Vec_Fetch(ivars->elems, ivars->tick);
+        String *name = (String*)Vec_Fetch(ivars->elems, (size_t)ivars->tick);
         if (name) {
             return CFReader_Local_Is_Directory(ivars->cf_reader, name);
         }
